+++ android/src/main/java/com/azihsoyn/flutter/mlkit/MlkitPlugin.java
@@ -28,7 +28,12 @@ import com.google.firebase.ml.vision.cloud.text.FirebaseVisionCloudText;
 import com.google.firebase.ml.vision.common.FirebaseVisionImage;
 import com.google.firebase.ml.vision.text.FirebaseVisionText;
 import com.google.firebase.ml.vision.common.FirebaseVisionImageMetadata;
-
+import com.google.firebase.ml.vision.face.FirebaseVisionFace;
+import com.google.firebase.ml.vision.face.FirebaseVisionFaceDetector;
+import com.google.firebase.ml.vision.face.FirebaseVisionFaceLandmark;
+import com.google.firebase.ml.vision.face.FirebaseVisionFaceDetectorOptions;
+import com.google.firebase.ml.vision.label.FirebaseVisionLabel;
+import com.google.firebase.ml.vision.label.FirebaseVisionLabelDetector;
 import android.graphics.Bitmap;
 import android.media.Image;
 import java.io.IOException;
@@ -40,12 +45,26 @@ import java.io.InputStream;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
+import java.util.Collections;
 
 /**
  * MlkitPlugin
  */
 public class MlkitPlugin implements MethodCallHandler {
   private static Context context;
+
+  private static final List<Integer> LandmarkTypes = Collections.unmodifiableList( new ArrayList<Integer>() {{
+    add(FirebaseVisionFaceLandmark.BOTTOM_MOUTH);
+    add(FirebaseVisionFaceLandmark.RIGHT_MOUTH);
+    add(FirebaseVisionFaceLandmark.LEFT_MOUTH);
+    add(FirebaseVisionFaceLandmark.RIGHT_EYE);
+    add(FirebaseVisionFaceLandmark.LEFT_EYE);
+    add(FirebaseVisionFaceLandmark.RIGHT_EAR);
+    add(FirebaseVisionFaceLandmark.LEFT_EAR);
+    add(FirebaseVisionFaceLandmark.RIGHT_CHEEK);
+    add(FirebaseVisionFaceLandmark.LEFT_CHEEK);
+    add(FirebaseVisionFaceLandmark.NOSE_BASE);
+  }} );
   /**
    * Plugin registration.
    */
@@ -109,6 +128,76 @@ public class MlkitPlugin implements MethodCallHandler {
                         });
       } catch (IOException e) {
         Log.e("error", e.getMessage());
+        return;
+      }
+    } else if (call.method.equals("FirebaseVisionLabelDetector#detectFromPath")){
+      String path = call.argument("filepath");
+      try {
+        File file = new File(path);
+        FirebaseVisionImage image = FirebaseVisionImage.fromFilePath(context, Uri.fromFile(file));
+        FirebaseVisionLabelDetector detector = FirebaseVision.getInstance()
+                .getVisionLabelDetector();
+        detector.detectInImage(image)
+                .addOnSuccessListener(
+                        new OnSuccessListener<List<FirebaseVisionLabel>>() {
+                          @Override
+                          public void onSuccess(List<FirebaseVisionLabel> labels) {
+                            result.success(processImageLabelingResult(labels));
+                          }
+                        })
+                .addOnFailureListener(
+                        new OnFailureListener() {
+                          @Override
+                          public void onFailure(@NonNull Exception e) {
+                            // Task failed with an exception
+                            e.printStackTrace();
+                          }
+                        });
+      }catch (IOException e){
+        Log.e("error",e.getMessage());
+        return;
+      }
+    } else if (call.method.equals("FirebaseVisionFaceDetector#detectFromPath")){
+      String path = call.argument("filepath");
+
+      try {
+        File file = new File(path);
+        FirebaseVisionImage image = FirebaseVisionImage.fromFilePath(context, Uri.fromFile(file));
+        FirebaseVisionFaceDetector detector;
+        if (call.argument("option") != null) {
+          Map<String, Object> optionsMap = call.argument("option");
+          FirebaseVisionFaceDetectorOptions options =
+                  new FirebaseVisionFaceDetectorOptions.Builder()
+                          .setModeType((int)optionsMap.get("modeType"))
+                          .setLandmarkType((int)optionsMap.get("landmarkType"))
+                          .setClassificationType((int)optionsMap.get("classificationType"))
+                          .setMinFaceSize((float)(double)optionsMap.get("minFaceSize"))
+                          .setTrackingEnabled((boolean)optionsMap.get("isTrackingEnabled"))
+                          .build();
+          detector = FirebaseVision.getInstance()
+                  .getVisionFaceDetector(options);
+        } else {
+          detector = FirebaseVision.getInstance()
+                  .getVisionFaceDetector();
+        }
+        detector.detectInImage(image)
+                .addOnSuccessListener(
+                        new OnSuccessListener<List<FirebaseVisionFace>>() {
+                          @Override
+                          public void onSuccess(List<FirebaseVisionFace> faces) {
+                            result.success(processFaceDetectionResult(faces));
+                          }
+                        })
+                .addOnFailureListener(
+                        new OnFailureListener() {
+                          @Override
+                          public void onFailure(@NonNull Exception e) {
+                            // Task failed with an exception
+                            e.printStackTrace();
+                          }
+                        });
+      }catch (IOException e){
+        Log.e("error",e.getMessage());
         return;
       }
     } else {
@@ -333,6 +422,64 @@ public class MlkitPlugin implements MethodCallHandler {
       blockBuilder.put("lines", linesBuilder.build());
       dataBuilder.add(blockBuilder.build());
     }
+    return dataBuilder.build();
+  }
+
+  private ImmutableList<ImmutableMap<String, Object>> processFaceDetectionResult(List<FirebaseVisionFace> faces) {
+    ImmutableList.Builder<ImmutableMap<String, Object>> dataBuilder =
+            ImmutableList.<ImmutableMap<String, Object>>builder();
+
+    for (FirebaseVisionFace face: faces) {
+      ImmutableMap.Builder<String, Object> faceBuilder = ImmutableMap.<String, Object>builder();
+      faceBuilder.put("rect_bottom", (double)face.getBoundingBox().bottom);
+      faceBuilder.put("rect_top", (double)face.getBoundingBox().top);
+      faceBuilder.put("rect_right", (double)face.getBoundingBox().right);
+      faceBuilder.put("rect_left", (double)face.getBoundingBox().left);
+      faceBuilder.put("tracking_id", (int)face.getTrackingId());
+      faceBuilder.put("head_euler_angle_y", face.getHeadEulerAngleY());
+      faceBuilder.put("head_euler_angle_z", face.getHeadEulerAngleZ());
+      faceBuilder.put("smiling_probability", face.getSmilingProbability());
+      faceBuilder.put("right_eye_open_probability", face.getRightEyeOpenProbability());
+      faceBuilder.put("left_eye_open_probability", face.getLeftEyeOpenProbability());
+      ImmutableMap.Builder<Integer, Object> landmarksBuilder = ImmutableMap.<Integer, Object>builder();
+      for (Integer landmarkType: LandmarkTypes) {
+        ImmutableMap.Builder<String, Object> landmarkBuilder = ImmutableMap.<String, Object>builder();
+        ImmutableMap.Builder<String, Object> positionkBuilder = ImmutableMap.<String, Object>builder();
+        FirebaseVisionFaceLandmark landmark = face.getLandmark(landmarkType);
+        if(landmark != null) {
+          positionkBuilder.put("x", landmark.getPosition().getX());
+          positionkBuilder.put("y", landmark.getPosition().getY());
+          if(landmark.getPosition().getZ() != null) {
+            positionkBuilder.put("z", landmark.getPosition().getZ());
+          }
+          landmarkBuilder.put("position", positionkBuilder.build());
+          landmarkBuilder.put("type", landmarkType);
+          landmarksBuilder.put(landmarkType, landmarkBuilder.build());
+        }
+      }
+      faceBuilder.put("landmarks", landmarksBuilder.build());
+
+
+      dataBuilder.add(faceBuilder.build());
+    }
+
+    return dataBuilder.build();
+  }
+
+  private ImmutableList<ImmutableMap<String, Object>> processImageLabelingResult(List<FirebaseVisionLabel> labels) {
+    ImmutableList.Builder<ImmutableMap<String, Object>> dataBuilder =
+            ImmutableList.<ImmutableMap<String, Object>>builder();
+
+    for (FirebaseVisionLabel label: labels) {
+      ImmutableMap.Builder<String, Object> labelBuilder = ImmutableMap.<String, Object>builder();
+
+      labelBuilder.put("label", label.getLabel());
+      labelBuilder.put("entityID", label.getEntityId());
+      labelBuilder.put("confidence", label.getConfidence());
+
+      dataBuilder.add(labelBuilder.build());
+    }
+
     return dataBuilder.build();
   }
 }