+++ CHANGELOG.md
@@ -1,4 +1,16 @@
-## 0.2.1
+## 0.4.1
+
+* Added On-device Face detecting API for android.
+
+## 0.4.0
+
+* Added On-device Face detecting API (iOS only).
+
+## 0.3.0
+
+* Added On-device Image Labeling API.
+
+## 0.2.2
 
 * fix package dependency for android.